Add files via upload
File Integrity Monitoring is designed to detect and alert when critical files or directories are modified, accessed, or corrupted, either intentionally or accidentally. The goal is to ensure that files crucial for security, compliance, or business operations remain in a known, trusted state.
A file intergrity was demonstrating using the english premier league table , a modification was carried to test the efficiency of wazuh in monitoring the original file modification. a complete xray of the file has been depicted and corresponding changes also highlighted and discussed.